python_backend/pyproject.toml
[project]
name = "python-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.2",
    "cloudinary>=1.44.1",
    "fastapi>=0.134.0",
    "google-generativeai>=0.8.6",
    "groq>=1.0.0",
    "langchain-text-splitters>=1.1.1",
    "loguru>=0.7.3",
    "motor>=3.7.1",
    "pgvector>=0.4.2",
    "prisma>=0.15.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.1",
    "python-multipart>=0.0.22",
    "redis>=7.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]

